+++ src/client/app/router.ts
@@ -1,9 +1,9 @@
-import { reactive, inject, markRaw, nextTick, readonly } from 'vue'
 import type { Component, InjectionKey } from 'vue'
+import { inject, markRaw, nextTick, reactive, readonly } from 'vue'
+import type { Awaitable, PageData, PageDataPayload } from '../shared'
 import { notFoundPageData, treatAsHtml } from '../shared'
-import type { PageData, PageDataPayload, Awaitable } from '../shared'
-import { inBrowser, withBase } from './utils'
 import { siteDataRef } from './data'
+import { getScrollOffset, inBrowser, withBase } from './utils'
 
 export interface Route {
   path: string
@@ -66,7 +66,11 @@ export function createRouter(
   async function go(href: string = inBrowser ? location.href : '/') {
     href = normalizeHref(href)
     if ((await router.onBeforeRouteChange?.(href)) === false) return
-    updateHistory(href)
+    if (inBrowser && href !== normalizeHref(location.href)) {
+      // save scroll position before changing url
+      history.replaceState({ scrollPosition: window.scrollY }, '')
+      history.pushState({}, '', href)
+    }
     await loadPage(href)
     await router.onAfterRouteChanged?.(href)
   }
@@ -107,7 +111,7 @@ export function createRouter(
             if (actualPathname !== targetLoc.pathname) {
               targetLoc.pathname = actualPathname
               href = actualPathname + targetLoc.search + targetLoc.hash
-              history.replaceState(null, '', href)
+              history.replaceState({}, '', href)
             }
 
             if (targetLoc.hash && !scrollPosition) {
@@ -158,6 +162,9 @@ export function createRouter(
   }
 
   if (inBrowser) {
+    if (history.state === null) {
+      history.replaceState({}, '')
+    }
     window.addEventListener(
       'click',
       (e) => {
@@ -180,7 +187,7 @@ export function createRouter(
               : link.href,
             link.baseURI
           )
-          const currentUrl = window.location
+          const currentUrl = new URL(location.href) // copy to keep old data
           // only intercept inbound html links
           if (
             !e.ctrlKey &&
@@ -199,15 +206,19 @@ export function createRouter(
               // scroll between hash anchors in the same page
               // avoid duplicate history entries when the hash is same
               if (hash !== currentUrl.hash) {
-                history.pushState(null, '', hash)
+                history.pushState({}, '', href)
                 // still emit the event so we can listen to it in themes
-                window.dispatchEvent(new Event('hashchange'))
+                window.dispatchEvent(
+                  new HashChangeEvent('hashchange', {
+                    oldURL: currentUrl.href,
+                    newURL: href
+                  })
+                )
               }
               if (hash) {
                 // use smooth scroll when clicking on header anchor links
                 scrollTo(link, hash, link.classList.contains('header-anchor'))
               } else {
-                updateHistory(href)
                 window.scrollTo(0, 0)
               }
             } else {
@@ -220,6 +231,9 @@ export function createRouter(
     )
 
     window.addEventListener('popstate', async (e) => {
+      if (e.state === null) {
+        return
+      }
       await loadPage(
         normalizeHref(location.href),
         (e.state && e.state.scrollPosition) || 0
@@ -261,26 +275,6 @@ export function scrollTo(el: Element, hash: string, smooth = false) {
   }
 
   if (target) {
-    let scrollOffset = siteDataRef.value.scrollOffset
-    let offset = 0
-    let padding = 24
-    if (typeof scrollOffset === 'object' && 'padding' in scrollOffset) {
-      padding = scrollOffset.padding
-      scrollOffset = scrollOffset.selector
-    }
-    if (typeof scrollOffset === 'number') {
-      offset = scrollOffset
-    } else if (typeof scrollOffset === 'string') {
-      offset = tryOffsetSelector(scrollOffset, padding)
-    } else if (Array.isArray(scrollOffset)) {
-      for (const selector of scrollOffset) {
-        const res = tryOffsetSelector(selector, padding)
-        if (res) {
-          offset = res
-          break
-        }
-      }
-    }
     const targetPadding = parseInt(
       window.getComputedStyle(target).paddingTop,
       10
@@ -288,7 +282,7 @@ export function scrollTo(el: Element, hash: string, smooth = false) {
     const targetTop =
       window.scrollY +
       target.getBoundingClientRect().top -
-      offset +
+      getScrollOffset() +
       targetPadding
     function scrollToTarget() {
       // only smooth scroll if distance is smaller than screen height.
@@ -298,14 +292,6 @@ export function scrollTo(el: Element, hash: string, smooth = false) {
     }
     requestAnimationFrame(scrollToTarget)
   }
-}
-
-function tryOffsetSelector(selector: string, padding: number): number {
-  const el = document.querySelector(selector)
-  if (!el) return 0
-  const bot = el.getBoundingClientRect().bottom
-  if (bot < 0) return 0
-  return bot + padding
 }
 
 function handleHMR(route: Route): void {
@@ -326,14 +312,6 @@ function shouldHotReload(payload: PageDataPayload): boolean {
     .replace(/(?:(^|\/)index)?\.html$/, '')
     .slice(siteDataRef.value.base.length - 1)
   return payloadPath === locationPath
-}
-
-function updateHistory(href: string) {
-  if (inBrowser && normalizeHref(href) !== normalizeHref(location.href)) {
-    // save scroll position before changing url
-    history.replaceState({ scrollPosition: window.scrollY }, document.title)
-    history.pushState(null, '', href)
-  }
 }
 
 function normalizeHref(href: string): string {